HEEDLS-692 Split SelfAssessmentService into data service and actual service
