+++ DigitalLearningSolutions.Data.Tests/DataServices/SelfAssessmentDataServiceTests/CompetencyDataServiceTests.cs
@@ -1,0 +1,387 @@
+﻿namespace DigitalLearningSolutions.Data.Tests.DataServices.SelfAssessmentDataServiceTests
+{
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Transactions;
+    using Dapper;
+    using DigitalLearningSolutions.Data.Models.SelfAssessments;
+    using DigitalLearningSolutions.Data.Tests.TestHelpers;
+    using FluentAssertions;
+    using NUnit.Framework;
+
+    public partial class SelfAssessmentDataServiceTests
+    {
+        [Test]
+        public void GetCompetencyIdsForSelfAssessment_returns_expected_ids()
+        {
+            // Given
+            var expectedIds = Enumerable.Range(1, 32).ToList();
+
+            // When
+            var result = selfAssessmentDataService.GetCompetencyIdsForSelfAssessment(1).ToList();
+
+            // Then
+            result.Should().BeEquivalentTo(expectedIds);
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_first_competency()
+        {
+            // Given
+            var expectedCompetency = SelfAssessmentHelper.CreateDefaultCompetency(
+                competencyGroup: "Data, information and content",
+                name: "I can find, use and store information that exists in different digital locations e.g. on a PC, shared drives, via the internet",
+                description: "I can find, use and store information that exists in different digital locations e.g. on a PC, shared drives, via the internet",
+                assessmentQuestions: new List<AssessmentQuestion>()
+                {
+                    SelfAssessmentHelper.CreateDefaultAssessmentQuestion(id: 1, question: "Where are you now"),
+                    SelfAssessmentHelper.CreateDefaultAssessmentQuestion(id: 2, question: "Where do you need to be")
+                }
+            );
+
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(1, SelfAssessmentId, CandidateId);
+
+            // Then
+            result.Should().BeEquivalentTo(expectedCompetency);
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_last_competency()
+        {
+            // Given
+            var expectedCompetency = SelfAssessmentHelper.CreateDefaultCompetency(
+                id: 32,
+                competencyGroup: "General questions",
+                competencyGroupId: 7,
+                name: "Taking an active role in my own learning is the most important thing that affects my digital literacy skills development",
+                description: "Taking an active role in my own learning is the most important thing that affects my digital literacy skills development",
+                assessmentQuestions: new List<AssessmentQuestion>()
+                {
+                    SelfAssessmentHelper.CreateDefaultAssessmentQuestion(
+                        id: 3,
+                        maxValueDescription: "Strongly agree",
+                        minValueDescription: "Strongly disagree",
+                        question: "To what extent do you agree",
+                       assessmentQuestionInputTypeID: 1
+                    )
+                });
+
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(32, SelfAssessmentId, CandidateId);
+
+            // Then
+            result.Should().BeEquivalentTo(expectedCompetency);
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_null_when_reached_end_of_assessment()
+        {
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(33, SelfAssessmentId, CandidateId);
+
+            // Then
+            result.Should().BeNull();
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_null_when_n_zero()
+        {
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(0, SelfAssessmentId, CandidateId);
+
+            // Then
+            result.Should().BeNull();
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_null_when_n_negative()
+        {
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(-1, SelfAssessmentId, CandidateId);
+
+            // Then
+            result.Should().BeNull();
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_null_when_invalid_candidate()
+        {
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(1, SelfAssessmentId, 1);
+
+            // Then
+            result.Should().BeNull();
+        }
+
+        [Test]
+        public void GetNthCompetency_returns_null_when_invalid_assessment()
+        {
+            // When
+            var result = selfAssessmentDataService.GetNthCompetency(1, 2, 1);
+
+            // Then
+            result.Should().BeNull();
+        }
+
+        [Test]
+        public void GetNthCompetency_gets_the_latest_result()
+        {
+            // Given
+            const int competencyId = 2;
+            const int assessmentQuestionId = 2;
+            const int result = 5;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, result + 1, null);
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, result, null);
+
+                // Then
+                var competency = selfAssessmentDataService.GetNthCompetency(2, SelfAssessmentId, CandidateId);
+                var actualResult = competency.AssessmentQuestions.First(question => question.Id == assessmentQuestionId).Result;
+                result.Should().Be(actualResult);
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_sets_result()
+        {
+            // Given
+            const int competencyId = 2;
+            const int assessmentQuestionId = 2;
+            const int result = 5;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, result, null);
+                var insertedResult = GetAssessmentResults(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId).First();
+
+                // Then
+                insertedResult.Should().Be(result);
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_overwrite_previous_result()
+        {
+            // Given
+            const int competencyId = 2;
+            const int assessmentQuestionId = 2;
+            const int firstResult = 5;
+            const int secondResult = 10;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, firstResult, null);
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, secondResult, null);
+                var insertedResults = GetAssessmentResults(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId).ToList();
+
+                // Then
+                insertedResults.Should().HaveCount(2);
+                insertedResults[0].Should().Be(firstResult);
+                insertedResults[1].Should().Be(secondResult);
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_set_result_for_invalid_candidate()
+        {
+            // Given
+            const int competencyId = 2;
+            const int assessmentQuestionId = 2;
+            const int result = 5;
+            const int invalidCandidateId = 1;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, invalidCandidateId, assessmentQuestionId, result, null);
+                var insertedResults = GetAssessmentResults(competencyId, SelfAssessmentId, invalidCandidateId, assessmentQuestionId);
+
+                // Then
+                insertedResults.Should().BeEmpty();
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_set_result_for_invalid_assessment()
+        {
+            // Given
+            const int competencyId = 2;
+            const int assessmentQuestionId = 2;
+            const int result = 5;
+            const int invalidSelfAssessmentId = 2;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, invalidSelfAssessmentId, CandidateId, assessmentQuestionId, result, null);
+                var insertedResults = GetAssessmentResults(competencyId, invalidSelfAssessmentId, CandidateId, assessmentQuestionId);
+
+                // Then
+                insertedResults.Should().BeEmpty();
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_set_result_for_invalid_competency()
+        {
+            // Given
+            const int invalidCompetencyId = 33;
+            const int assessmentQuestionId = 2;
+            const int result = 5;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(invalidCompetencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, result, null);
+                var insertedResults = GetAssessmentResults(invalidCompetencyId, SelfAssessmentId, CandidateId, assessmentQuestionId);
+
+                // Then
+                insertedResults.Should().BeEmpty();
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_set_result_for_invalid_question()
+        {
+            // Given
+            const int competencyId = 33;
+            const int invalidAssessmentQuestionId = 4;
+            const int result = 5;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, invalidAssessmentQuestionId, result, null);
+                var insertedResults = GetAssessmentResults(competencyId, SelfAssessmentId, CandidateId, invalidAssessmentQuestionId);
+
+                // Then
+                insertedResults.Should().BeEmpty();
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_set_result_for_negative_result()
+        {
+            // Given
+            const int competencyId = 33;
+            const int assessmentQuestionId = 4;
+            const int invalidResult = -1;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, invalidResult, null);
+                var insertedResults = GetAssessmentResults(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId);
+
+                // Then
+                insertedResults.Should().BeEmpty();
+            }
+        }
+
+        [Test]
+        public void SetResultForCompetency_does_not_set_result_for_invalid_result()
+        {
+            // Given
+            const int competencyId = 33;
+            const int assessmentQuestionId = 4;
+            const int invalidResult = 11;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId, invalidResult, null);
+                var insertedResults = GetAssessmentResults(competencyId, SelfAssessmentId, CandidateId, assessmentQuestionId);
+
+                // Then
+                insertedResults.Should().BeEmpty();
+            }
+        }
+
+        [Test]
+        public void GetMostRecentResults_gets_multiple_competencies()
+        {
+            // Given
+            const int firstCompetencyId = 2;
+            const int secondCompetencyId = 3;
+            const int firstAssessmentQuestionId = 1;
+            const int secondAssessmentQuestionId = 2;
+            const int thirdAssessmentQuestionId = 1;
+            const int fourthAssessmentQuestionId = 2;
+            const int firstResult = 1;
+            const int secondResult = 2;
+            const int thirdResult = 3;
+            const int fourthResult = 4;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(firstCompetencyId, SelfAssessmentId, CandidateId, firstAssessmentQuestionId, firstResult, null);
+                selfAssessmentDataService.SetResultForCompetency(firstCompetencyId, SelfAssessmentId, CandidateId, secondAssessmentQuestionId, secondResult, null);
+                selfAssessmentDataService.SetResultForCompetency(secondCompetencyId, SelfAssessmentId, CandidateId, thirdAssessmentQuestionId, thirdResult, null);
+                selfAssessmentDataService.SetResultForCompetency(secondCompetencyId, SelfAssessmentId, CandidateId, fourthAssessmentQuestionId, fourthResult, null);
+
+                // Then
+                var results = selfAssessmentDataService.GetMostRecentResults(SelfAssessmentId, CandidateId).ToList();
+
+                results.Count.Should().Be(32);
+                SelfAssessmentHelper.GetQuestionResult(results, firstCompetencyId, firstAssessmentQuestionId).Should().Be(firstResult);
+                SelfAssessmentHelper.GetQuestionResult(results, firstCompetencyId, secondAssessmentQuestionId).Should().Be(secondResult);
+                SelfAssessmentHelper.GetQuestionResult(results, secondCompetencyId, thirdAssessmentQuestionId).Should().Be(thirdResult);
+                SelfAssessmentHelper.GetQuestionResult(results, secondCompetencyId, fourthAssessmentQuestionId).Should().Be(fourthResult);
+                SelfAssessmentHelper.GetQuestionResult(results, 4, 1).Should().BeNull();
+            }
+        }
+
+        [Test]
+        public void GetMostRecentResults_gets_most_recent_results()
+        {
+            // Given
+            const int firstCompetencyId = 2;
+            const int secondCompetencyId = 3;
+            const int firstAssessmentQuestionId = 1;
+            const int secondAssessmentQuestionId = 2;
+            const int thirdAssessmentQuestionId = 1;
+            const int fourthAssessmentQuestionId = 2;
+            const int firstResult = 1;
+            const int secondResult = 2;
+            const int thirdResult = 3;
+            const int fourthResult = 4;
+
+            using (new TransactionScope())
+            {
+                // When
+                selfAssessmentDataService.SetResultForCompetency(firstCompetencyId, SelfAssessmentId, CandidateId, firstAssessmentQuestionId, firstResult, null);
+                selfAssessmentDataService.SetResultForCompetency(firstCompetencyId, SelfAssessmentId, CandidateId, secondAssessmentQuestionId, secondResult, null);
+                selfAssessmentDataService.SetResultForCompetency(secondCompetencyId, SelfAssessmentId, CandidateId, thirdAssessmentQuestionId, 9, null);
+                selfAssessmentDataService.SetResultForCompetency(secondCompetencyId, SelfAssessmentId, CandidateId, fourthAssessmentQuestionId, 9, null);
+                selfAssessmentDataService.SetResultForCompetency(secondCompetencyId, SelfAssessmentId, CandidateId, thirdAssessmentQuestionId, thirdResult, null);
+                selfAssessmentDataService.SetResultForCompetency(secondCompetencyId, SelfAssessmentId, CandidateId, fourthAssessmentQuestionId, fourthResult, null);
+
+                // Then
+                var results = selfAssessmentDataService.GetMostRecentResults(SelfAssessmentId, CandidateId).ToList();
+
+                results.Count.Should().Be(32);
+                SelfAssessmentHelper.GetQuestionResult(results, secondCompetencyId, thirdAssessmentQuestionId).Should().Be(thirdResult);
+                SelfAssessmentHelper.GetQuestionResult(results, secondCompetencyId, fourthAssessmentQuestionId).Should().Be(fourthResult);
+            }
+        }
+
+        private IEnumerable<int> GetAssessmentResults(int competencyId, int selfAssessmentId, int candidateId, int assessmentQuestionId)
+        {
+            return connection.Query<int>(
+                @"SELECT Result FROM SelfAssessmentResults WHERE
+                        CompetencyID = @competencyId AND
+                        SelfAssessmentID = @selfAssessmentId AND
+                        CandidateID = @candidateId AND
+                        AssessmentQuestionID = @assessmentQuestionId",
+                new { competencyId, selfAssessmentId, candidateId, assessmentQuestionId }
+            );
+        }
+    }
+}
