+++ DigitalLearningSolutions.Data/DataServices/SelfAssessmentDataService/SelfAssessmentDataService.cs
@@ -1,0 +1,130 @@
+﻿namespace DigitalLearningSolutions.Data.DataServices.SelfAssessmentDataService
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Data;
+    using DigitalLearningSolutions.Data.Models.Common.Users;
+    using DigitalLearningSolutions.Data.Models.External.Filtered;
+    using DigitalLearningSolutions.Data.Models.Frameworks;
+    using DigitalLearningSolutions.Data.Models.SelfAssessments;
+    using Microsoft.Extensions.Logging;
+
+    public interface ISelfAssessmentDataService
+    {
+        // CompetencyDataService
+        IEnumerable<int> GetCompetencyIdsForSelfAssessment(int selfAssessmentId);
+
+        Competency? GetNthCompetency(int n, int selfAssessmentId, int candidateId); // 1 indexed
+
+        IEnumerable<Competency> GetMostRecentResults(int selfAssessmentId, int candidateId);
+
+        IEnumerable<Competency> GetCandidateAssessmentResultsById(int candidateAssessmentId, int adminId);
+
+        IEnumerable<Competency> GetCandidateAssessmentResultsForReviewById(int candidateAssessmentId, int adminId);
+
+        IEnumerable<Competency> GetCandidateAssessmentResultsToVerifyById(int selfAssessmentId, int candidateId);
+
+        Competency? GetCompetencyByCandidateAssessmentResultId(int resultId, int candidateAssessmentId, int adminId);
+
+        void SetResultForCompetency(
+            int competencyId,
+            int selfAssessmentId,
+            int candidateId,
+            int assessmentQuestionId,
+            int? result,
+            string? supportingComments
+        );
+
+        IEnumerable<Competency> GetCandidateAssessmentOptionalCompetencies(int selfAssessmentId, int candidateId);
+
+        void InsertCandidateAssessmentOptionalCompetenciesIfNotExist(int selfAssessmentId, int candidateId);
+
+        void UpdateCandidateAssessmentOptionalCompetencies(int selfAssessmentStructureId, int candidateId);
+
+        IEnumerable<LevelDescriptor> GetLevelDescriptorsForAssessmentQuestion(
+            int assessmentQuestionId,
+            int minValue,
+            int maxValue,
+            bool zeroBased
+        );
+
+        List<int> GetCandidateAssessmentIncludedSelfAssessmentStructureIds(int selfAssessmentId, int candidateId);
+
+        // CandidateAssessmentsDataService
+        IEnumerable<CurrentSelfAssessment> GetSelfAssessmentsForCandidate(int candidateId);
+
+        CurrentSelfAssessment? GetSelfAssessmentForCandidateById(int candidateId, int selfAssessmentId);
+
+        void UpdateLastAccessed(int selfAssessmentId, int candidateId);
+
+        void SetCompleteByDate(int selfAssessmentId, int candidateId, DateTime? completeByDate);
+
+        void SetSubmittedDateNow(int selfAssessmentId, int candidateId);
+
+        void IncrementLaunchCount(int selfAssessmentId, int candidateId);
+
+        void SetUpdatedFlag(int selfAssessmentId, int candidateId, bool status);
+
+        void SetBookmark(int selfAssessmentId, int candidateId, string bookmark);
+
+        // SelfAssessmentSupervisorDataService
+        SelfAssessmentSupervisor? GetSupervisorForSelfAssessmentId(int selfAssessmentId, int candidateId);
+
+        IEnumerable<SelfAssessmentSupervisor> GetSupervisorsForSelfAssessmentId(int selfAssessmentId, int candidateId);
+
+        IEnumerable<SelfAssessmentSupervisor> GetOtherSupervisorsForCandidate(int selfAssessmentId, int candidateId);
+
+        IEnumerable<SelfAssessmentSupervisor> GetAllSupervisorsForSelfAssessmentId(
+            int selfAssessmentId,
+            int candidateId
+        );
+
+        IEnumerable<SelfAssessmentSupervisor> GetResultReviewSupervisorsForSelfAssessmentId(
+            int selfAssessmentId,
+            int candidateId
+        );
+
+        IEnumerable<SelfAssessmentSupervisor> GetSignOffSupervisorsForSelfAssessmentId(
+            int selfAssessmentId,
+            int candidateId
+        );
+
+        SelfAssessmentSupervisor? GetSelfAssessmentSupervisorByCandidateAssessmentSupervisorId(
+            int candidateAssessmentSupervisorId
+        );
+
+        void InsertCandidateAssessmentSupervisorVerification(int candidateAssessmentSupervisorId);
+
+        void UpdateCandidateAssessmentSupervisorVerificationEmailSent(int candidateAssessmentSupervisorVerificationId);
+
+        SupervisorComment? GetSupervisorComments(int candidateId, int resultId);
+
+        IEnumerable<Administrator> GetValidSupervisorsForActivity(int centreId, int selfAssessmentId, int candidateId);
+
+        Administrator GetSupervisorByAdminId(int supervisorAdminId);
+
+        IEnumerable<SupervisorSignOff> GetSupervisorSignOffsForCandidateAssessment(
+            int selfAssessmentId,
+            int candidateId
+        );
+
+        // FilteredDataService
+        Profile? GetFilteredProfileForCandidateById(int candidateId, int selfAssessmentId);
+
+        IEnumerable<Goal> GetFilteredGoalsForCandidateId(int candidateId, int selfAssessmentId);
+
+        void LogAssetLaunch(int candidateId, int selfAssessmentId, LearningAsset learningAsset);
+    }
+
+    public partial class SelfAssessmentDataService : ISelfAssessmentDataService
+    {
+        private readonly IDbConnection connection;
+        private readonly ILogger<SelfAssessmentDataService> logger;
+
+        public SelfAssessmentDataService(IDbConnection connection, ILogger<SelfAssessmentDataService> logger)
+        {
+            this.connection = connection;
+            this.logger = logger;
+        }
+    }
+}
